requirements.txt
# Core ML libraries
tensorflow==2.20.0
numpy==1.24.3
pandas==2.1.0

# Text processing
python-levenshtein==0.21.1
pyarabic==0.6.15

# Configuration
pyyaml==6.0.1

# Visualization and analysis
matplotlib==3.7.2
seaborn==0.12.2

# Development
pytest==7.4.0
jupyter==1.0.0
notebook==7.0.2

# Utilities
tqdm==4.66.1
scikit-learn==1.3.0